Implemented mouse shift and rotate
Implemented shifting and rotating of structures with the mouse. This
works by holding 's' or 'r', respectively, and dragging with the mouse.

Note that the operations only affect the active atoms. Also note that
rotating rotates particles in place. That is, atoms are not affected -
only clusters rotate around their centers.